requirements.txt
# Core ML
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
mplfinance>=0.12.10b0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# Web UI
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
websockets>=12.0

# LLM Providers (install as needed)
openai>=1.30.0
anthropic>=0.40.0
google-generativeai>=0.8.0
